+++ samples/sampler/stories/default/FormCheckboxItem.js
@@ -1,6 +1,6 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
+import {boolean, select, text} from '@enact/storybook-utils/addons/controls';
 import Checkbox, {CheckboxBase} from '@enact/sandstone/Checkbox';
 import FormCheckboxItem, {FormCheckboxItemBase} from '@enact/sandstone/FormCheckboxItem';
 import Icon from '@enact/sandstone/Icon';
@@ -25,28 +25,38 @@ export default {
 	component: 'FormCheckboxItem'
 };
 
-export const _FormCheckboxItem = () => {
-	const slotBeforeSelection = select('slotBefore', ['', ...iconNames], Config);
+export const _FormCheckboxItem = (args) => {
+	const slotBeforeSelection = args['slotBefore'];
 	const slotBefore = slotBeforeSelection ? (
 		<Icon slot="slotBefore">{slotBeforeSelection}</Icon>
 	) : null;
 
 	return (
 		<FormCheckboxItem
-			disabled={boolean('disabled', Config)}
-			icon={select('icon', iconNames, Config)}
-			indeterminate={boolean('indeterminate', Config)}
-			indeterminateIcon={select('indeterminateIcon', iconNames, Config)}
-			inline={boolean('inline', Config)}
-			label={text('label', Config)}
-			labelPosition={select('labelPosition', ['above', 'after', 'before', 'below'], Config)}
+			disabled={args['disabled']}
+			icon={args['icon']}
+			indeterminate={args['indeterminate']}
+			indeterminateIcon={args['indeterminateIcon']}
+			inline={args['inline']}
+			label={args['label']}
+			labelPosition={args['labelPosition']}
 			onToggle={action('onToggle')}
 		>
 			{slotBefore}
-			{text('children', Config, 'A Checkbox for a form')}
+			{args['children']}
 		</FormCheckboxItem>
 	);
 };
+
+select('slotBefore', _FormCheckboxItem, ['', ...iconNames], Config);
+boolean('disabled', _FormCheckboxItem, Config);
+select('icon', _FormCheckboxItem, iconNames, Config);
+boolean('indeterminate', _FormCheckboxItem, Config);
+select('indeterminateIcon', _FormCheckboxItem, iconNames, Config);
+boolean('inline', _FormCheckboxItem, Config);
+text('label', _FormCheckboxItem, Config);
+select('labelPosition', _FormCheckboxItem, ['above', 'after', 'before', 'below'], Config);
+text('children', _FormCheckboxItem, Config, 'A Checkbox for a form');
 
 _FormCheckboxItem.storyName = 'FormCheckboxItem';
 _FormCheckboxItem.parameters = {