+++ samples/sampler/stories/default/SwitchItem.js
@@ -1,6 +1,6 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
+import {boolean, select, text} from '@enact/storybook-utils/addons/controls';
 import Icon from '@enact/sandstone/Icon';
 import Item, {ItemBase} from '@enact/sandstone/Item';
 import SwitchItem from '@enact/sandstone/SwitchItem';
@@ -15,21 +15,26 @@ export default {
 	component: 'SwitchItem'
 };
 
-export const _SwitchItem = () => {
-	const slotAfterSelection = select('slotAfter', ['', ...iconNames], Config);
+export const _SwitchItem = (args) => {
+	const slotAfterSelection = args['slotAfter'];
 	const slotAfter = slotAfterSelection ? <Icon slot="slotAfter">{slotAfterSelection}</Icon> : null;
 
 	return (
 		<SwitchItem
-			disabled={boolean('disabled', Config)}
-			inline={boolean('inline', Config)}
+			disabled={args['disabled']}
+			inline={args['inline']}
 			onToggle={action('onToggle')}
 		>
-			{text('children', Config, 'Hello SwitchItem')}
+			{args['children']}
 			{slotAfter}
 		</SwitchItem>
 	);
 };
+
+select('slotAfter', _SwitchItem, ['', ...iconNames], Config);
+boolean('disabled', _SwitchItem, Config);
+boolean('inline', _SwitchItem, Config);
+text('children', _SwitchItem, Config, 'Hello SwitchItem');
 
 _SwitchItem.storyName = 'SwitchItem';
 _SwitchItem.parameters = {