+++ samples/sampler/stories/qa/Scroller.js
@@ -1,7 +1,8 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
+import {boolean, range, select} from '@enact/storybook-utils/addons/controls';
 import Button from '@enact/sandstone/Button';
+import BodyText from '@enact/sandstone/BodyText';
 import Item from '@enact/sandstone/Item';
 import Scroller from '@enact/sandstone/Scroller';
 import Spotlight from '@enact/spotlight';
@@ -26,7 +27,7 @@ const prop = {
 	focusableScrollbarOption: {
 		false: false,
 		true: true,
-		'&quot;byEnter&quot;': 'byEnter'
+		byEnter: 'byEnter'
 	},
 	scrollbarOption: ['auto', 'hidden', 'visible'],
 	scrollModeOption: ['native', 'translate']
@@ -72,18 +73,15 @@ class ScrollerWithLongItem extends Component {
 	};
 
 	render () {
+		const args = this.props.args;
 		return (
 			<Scroller
-				focusableScrollbar={
-					prop.focusableScrollbarOption[
-						select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-					]
-				}
-				key={select('scrollMode', prop.scrollModeOption, Config)}
+				focusableScrollbar={args['focusableScrollbar']}
+				key={args['scrollMode']}
 				onKeyDown={action('onKeyDown')}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+				scrollMode={args['scrollMode']}
 			>
 				<Item>
 					Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long Long
@@ -108,13 +106,14 @@ class ScrollerWithResizable extends Component {
 	};
 
 	render () {
+		const args = this.props.args;
 		return (
 			<Scroller
-				key={select('scrollMode', prop.scrollModeOption, Config)}
+				key={args['scrollMode']}
 				onKeyDown={action('onKeyDown')}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+				scrollMode={args['scrollMode']}
 				verticalScrollbar="visible"
 			>
 				<Item>Item</Item>
@@ -135,18 +134,15 @@ class ScrollerWithLargeContainer extends Component {
 	}
 
 	render () {
+		const args = this.props.args;
 		return (
 			<Scroller
-				focusableScrollbar={
-					prop.focusableScrollbarOption[
-						select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-					]
-				}
-				key={select('scrollMode', prop.scrollModeOption, Config)}
+				focusableScrollbar={args['focusableScrollbar']}
+				key={args['scrollMode']}
 				onKeyDown={action('onKeyDown')}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+				scrollMode={args['scrollMode']}
 				spotlightId="scroller"
 				style={{height: ri.scaleToRem(600)}}
 			>
@@ -171,48 +167,48 @@ export default {
 	component: 'Scroller'
 };
 
-export const ListOfThings = () => (
+export const ListOfThings = (args) => (
 	<Scroller
-		focusableScrollbar={
-			prop.focusableScrollbarOption[
-				select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-			]
-		}
-		horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
-		hoverToScroll={boolean('hoverToScroll', Config)}
-		key={select('scrollMode', prop.scrollModeOption, Config)}
-		noScrollByWheel={boolean('noScrollByWheel', Config)}
+		focusableScrollbar={args['focusableScrollbar']}
+		horizontalScrollbar={args['horizontalScrollbar']}
+		hoverToScroll={args['hoverToScroll']}
+		key={args['scrollMode']}
+		noScrollByWheel={args['noScrollByWheel']}
 		onKeyDown={action('onKeyDown')}
 		onScrollStart={action('onScrollStart')}
 		onScrollStop={action('onScrollStop')}
-		scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
-		spotlightDisabled={boolean('spotlightDisabled', Config, false)}
-		verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
+		scrollMode={args['scrollMode']}
+		spotlightDisabled={args['spotlightDisabled']}
+		verticalScrollbar={args['verticalScrollbar']}
 	>
 		<Group childComponent={Item}>{itemData}</Group>
 	</Scroller>
 );
 
+select('focusableScrollbar', ListOfThings, prop.focusableScrollbarOption, Config);
+select('horizontalScrollbar', ListOfThings, prop.scrollbarOption, Config);
+boolean('hoverToScroll', ListOfThings, Config);
+boolean('noScrollByWheel', ListOfThings, Config);
+select('scrollMode', ListOfThings, prop.scrollModeOption, Config);
+boolean('spotlightDisabled', ListOfThings, Config, false);
+select('verticalScrollbar', ListOfThings, prop.scrollbarOption, Config);
+
 ListOfThings.storyName = 'List of things';
 
-export const HorizontalScroll = () => (
+export const HorizontalScroll = (args) => (
 	<Scroller
-		direction={select('direction', prop.direction, Config, 'horizontal')}
-		focusableScrollbar={
-			prop.focusableScrollbarOption[
-				select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-			]
-		}
-		horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
-		hoverToScroll={boolean('hoverToScroll', Config)}
-		key={select('scrollMode', prop.scrollModeOption, Config)}
-		noScrollByWheel={boolean('noScrollByWheel', Config)}
+		direction={args['direction']}
+		focusableScrollbar={args['focusableScrollbar']}
+		horizontalScrollbar={args['horizontalScrollbar']}
+		hoverToScroll={args['hoverToScroll']}
+		key={args['scrollMode']}
+		noScrollByWheel={args['noScrollByWheel']}
 		onKeyDown={action('onKeyDown')}
 		onScrollStart={action('onScrollStart')}
 		onScrollStop={action('onScrollStop')}
-		scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
-		spotlightDisabled={boolean('spotlightDisabled', Config, false)}
-		verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
+		scrollMode={args['scrollMode']}
+		spotlightDisabled={args['spotlightDisabled']}
+		verticalScrollbar={args['verticalScrollbar']}
 	>
 		<div
 			style={{
@@ -227,26 +223,31 @@ export const HorizontalScroll = () => (
 	</Scroller>
 );
 
+select('direction', HorizontalScroll, prop.direction, Config, 'horizontal');
+select('focusableScrollbar', HorizontalScroll, prop.focusableScrollbarOption, Config);
+select('horizontalScrollbar', HorizontalScroll, prop.scrollbarOption, Config);
+boolean('hoverToScroll', HorizontalScroll, Config);
+boolean('noScrollByWheel', HorizontalScroll, Config);
+select('scrollMode', HorizontalScroll, prop.scrollModeOption, Config);
+boolean('spotlightDisabled', HorizontalScroll, Config, false);
+select('verticalScrollbar', HorizontalScroll, prop.scrollbarOption, Config);
+
 HorizontalScroll.storyName = 'Horizontal scroll';
 
-export const WithSpottableComponents = () => (
+export const WithSpottableComponents = (args) => (
 	<Scroller
-		direction={select('direction', prop.direction, Config)}
-		focusableScrollbar={
-			prop.focusableScrollbarOption[
-				select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-			]
-		}
-		horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
-		hoverToScroll={boolean('hoverToScroll', Config)}
-		key={select('scrollMode', prop.scrollModeOption, Config)}
-		noScrollByWheel={boolean('noScrollByWheel', Config)}
+		direction={args['direction']}
+		focusableScrollbar={args['focusableScrollbar']}
+		horizontalScrollbar={args['horizontalScrollbar']}
+		hoverToScroll={args['hoverToScroll']}
+		key={args['scrollMode']}
+		noScrollByWheel={args['noScrollByWheel']}
 		onKeyDown={action('onKeyDown')}
 		onScrollStart={action('onScrollStart')}
 		onScrollStop={action('onScrollStop')}
-		scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
-		spotlightDisabled={boolean('spotlightDisabled', Config, false)}
-		verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
+		scrollMode={args['scrollMode']}
+		spotlightDisabled={args['spotlightDisabled']}
+		verticalScrollbar={args['verticalScrollbar']}
 	>
 		<div
 			style={{
@@ -275,27 +276,32 @@ export const WithSpottableComponents = () => (
 	</Scroller>
 );
 
+select('direction', WithSpottableComponents, prop.direction, Config);
+select('focusableScrollbar', WithSpottableComponents, prop.focusableScrollbarOption, Config);
+select('horizontalScrollbar', WithSpottableComponents, prop.scrollbarOption, Config);
+boolean('hoverToScroll', WithSpottableComponents, Config);
+boolean('noScrollByWheel', WithSpottableComponents, Config);
+select('scrollMode', WithSpottableComponents, prop.scrollModeOption, Config);
+boolean('spotlightDisabled', WithSpottableComponents, Config, false);
+select('verticalScrollbar', WithSpottableComponents, prop.scrollbarOption, Config);
+
 WithSpottableComponents.storyName = 'With Spottable Components';
 
-export const WithShortContents = () => (
+export const WithShortContents = (args) => (
 	<>
 		<Scroller
-			direction={select('direction', prop.direction, Config)}
-			focusableScrollbar={
-				prop.focusableScrollbarOption[
-					select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-				]
-			}
-			horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, Config)}
-			key={select('scrollMode', prop.scrollModeOption, Config)}
-			noScrollByWheel={boolean('noScrollByWheel', Config)}
+			direction={args['direction']}
+			focusableScrollbar={args['focusableScrollbar']}
+			horizontalScrollbar={args['horizontalScrollbar']}
+			key={args['scrollMode']}
+			noScrollByWheel={args['noScrollByWheel']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
-			spotlightDisabled={boolean('spotlightDisabled', Config, false)}
+			scrollMode={args['scrollMode']}
+			spotlightDisabled={args['spotlightDisabled']}
 			style={{height: ri.scaleToRem(600)}}
-			verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, Config)}
+			verticalScrollbar={args['verticalScrollbar']}
 		>
 			Text
 		</Scroller>
@@ -303,55 +309,75 @@ export const WithShortContents = () => (
 	</>
 );
 
+select('direction', WithShortContents, prop.direction, Config);
+select('focusableScrollbar', WithShortContents, prop.focusableScrollbarOption, Config);
+select('horizontalScrollbar', WithShortContents, prop.scrollbarOption, Config);
+boolean('hoverToScroll', WithShortContents, Config);
+boolean('noScrollByWheel', WithShortContents, Config);
+select('scrollMode', WithShortContents, prop.scrollModeOption, Config);
+boolean('spotlightDisabled', WithShortContents, Config, false);
+select('verticalScrollbar', WithShortContents, prop.scrollbarOption, Config);
+
 WithShortContents.storyName = 'With short contents';
 
-export const WithResizable = () => <ScrollerWithResizable />;
+ScrollerWithResizable.propTypes = {
+	args: PropTypes.object
+};
+
+export const WithResizable = (args) => <ScrollerWithResizable args={args} />;
+
+select('scrollMode', WithResizable, prop.scrollModeOption, Config);
 
 WithResizable.storyName = 'With Resizable';
 
-export const WithTwoUiScroller = () => (
+export const WithTwoUiScroller = (args) => (
 	<div style={{display: 'flex', height: ri.scaleToRem(798)}}>
 		<UiScroller
-			key={select('scrollMode', prop.scrollModeOption, Config) + '1'}
+			key={args['scrollMode'] + '1'}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 		>
 			<Group childComponent={Item}>{itemData}</Group>
 		</UiScroller>
 		<UiScroller
-			key={select('scrollMode', prop.scrollModeOption, Config) + '2'}
+			key={args['scrollMode'] + '2'}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 		>
 			<Group childComponent={Item}>{itemData}</Group>
 		</UiScroller>
 	</div>
 );
 
+select('scrollMode', WithTwoUiScroller, prop.scrollModeOption, Config);
+
 WithTwoUiScroller.storyName = 'With Two ui:Scroller';
 
-export const WithLargeContainer = () => <ScrollerWithLargeContainer />;
+ScrollerWithLargeContainer.propTypes = {
+	args: PropTypes.object
+};
+
+export const WithLargeContainer = (args) => <ScrollerWithLargeContainer args={args} />;
+
+select('focusableScrollbar', WithLargeContainer, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithLargeContainer, prop.scrollModeOption, Config);
 
 WithLargeContainer.storyName = 'With Large Container';
 
-export const WithFocusOutsideContainer = () => (
+export const WithFocusOutsideContainer = (args) => (
 	<div style={{display: 'flex'}}>
 		<Button>focus to me</Button>
 		<Scroller
-			focusableScrollbar={
-				prop.focusableScrollbarOption[
-					select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-				]
-			}
-			key={select('scrollMode', prop.scrollModeOption, Config)}
+			focusableScrollbar={args['focusableScrollbar']}
+			key={args['scrollMode']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 			style={{height: ri.scaleToRem(840), width: ri.scaleToRem(600), display: 'inline-block'}}
 		>
 			<Item>Item 1</Item>
@@ -368,17 +394,20 @@ export const WithFocusOutsideContainer = () => (
 	</div>
 );
 
+select('focusableScrollbar', WithFocusOutsideContainer, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithFocusOutsideContainer, prop.scrollModeOption, Config);
+
 WithFocusOutsideContainer.storyName = 'With Focus outside Container';
 
-export const TestScrollingToBoundaryWithSmallOverflow = () => {
-	const size = number('Spacer size', Config, {max: 600, min: 0, range: true}, 200);
+export const TestScrollingToBoundaryWithSmallOverflow = (args) => {
+	const size = args['Spacer size'];
 	return (
 		<Scroller
-			key={select('scrollMode', prop.scrollModeOption, Config)}
+			key={args['scrollMode']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 			style={{height: ri.scaleToRem(480)}}
 		>
 			<Item>1</Item>
@@ -390,22 +419,21 @@ export const TestScrollingToBoundaryWithSmallOverflow = () => {
 	);
 };
 
+range('Spacer size', TestScrollingToBoundaryWithSmallOverflow, Config, {max: 600, min: 0}, 200);
+select('scrollMode', TestScrollingToBoundaryWithSmallOverflow, prop.scrollModeOption, Config);
+
 TestScrollingToBoundaryWithSmallOverflow.storyName = 'Test scrolling to boundary with small overflow';
 
-export const TestScrollingToBoundaryWithLongOverflow = () => {
-	const size = number('Spacer size', Config, {max: 600, min: 0, range: true}, 402);
+export const TestScrollingToBoundaryWithLongOverflow = (args) => {
+	const size = args['Spacer size'];
 	return (
 		<Scroller
-			focusableScrollbar={
-				prop.focusableScrollbarOption[
-					select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-				]
-			}
-			key={select('scrollMode', prop.scrollModeOption, Config)}
+			focusableScrollbar={args['focusableScrollbar']}
+			key={args['scrollMode']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 			style={{height: ri.scaleToRem(402)}}
 		>
 			<div style={{height: ri.scaleToRem(size), paddingLeft: ri.scaleToRem(80)}}>
@@ -423,22 +451,22 @@ export const TestScrollingToBoundaryWithLongOverflow = () => {
 	);
 };
 
+range('Spacer size', TestScrollingToBoundaryWithLongOverflow, Config, {max: 600, min: 0}, 402);
+select('focusableScrollbar', TestScrollingToBoundaryWithLongOverflow, prop.focusableScrollbarOption, Config);
+select('scrollMode', TestScrollingToBoundaryWithLongOverflow, prop.scrollModeOption, Config);
+
 TestScrollingToBoundaryWithLongOverflow.storyName = 'Test scrolling to boundary with long overflow';
 
-export const WithSpotlightTargetCalculation = () => (
+export const WithSpotlightTargetCalculation = (args) => (
 	<div>
 		<Button>hello</Button>
 		<Scroller
-			focusableScrollbar={
-				prop.focusableScrollbarOption[
-					select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-				]
-			}
-			key={select('scrollMode', prop.scrollModeOption, Config)}
+			focusableScrollbar={args['focusableScrollbar']}
+			key={args['scrollMode']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 			style={{height: ri.scaleToRem(804)}}
 		>
 			<Group childComponent={Item}>{itemData}</Group>
@@ -446,24 +474,30 @@ export const WithSpotlightTargetCalculation = () => (
 	</div>
 );
 
+select('focusableScrollbar', WithSpotlightTargetCalculation, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithSpotlightTargetCalculation, prop.scrollModeOption, Config);
+
 WithSpotlightTargetCalculation.storyName = 'With Spotlight Target Calculation';
 
-export const WithLongItem = () => <ScrollerWithLongItem />;
+ScrollerWithLongItem.propTypes = {
+	args: PropTypes.object
+};
+
+export const WithLongItem = (args) => <ScrollerWithLongItem args={args} />;
+
+select('focusableScrollbar', WithLongItem, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithLongItem, prop.scrollModeOption, Config);
 
 WithLongItem.storyName = 'With Long Item';
 
-export const WithOneLongHeightItem = () => (
+export const WithOneLongHeightItem = (args) => (
 	<Scroller
-		focusableScrollbar={
-			prop.focusableScrollbarOption[
-				select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-			]
-		}
-		key={select('scrollMode', prop.scrollModeOption, Config)}
+		focusableScrollbar={args['focusableScrollbar']}
+		key={args['scrollMode']}
 		onKeyDown={action('onKeyDown')}
 		onScrollStart={action('onScrollStart')}
 		onScrollStop={action('onScrollStop')}
-		scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+		scrollMode={args['scrollMode']}
 	>
 		<div style={{height: ri.scaleToRem(2442)}}>
 			<Item style={{height: ri.scaleToRem(2400)}}>Long Height Item</Item>
@@ -471,39 +505,31 @@ export const WithOneLongHeightItem = () => (
 	</Scroller>
 );
 
+select('focusableScrollbar', WithOneLongHeightItem, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithOneLongHeightItem, prop.scrollModeOption, Config);
+
 WithOneLongHeightItem.storyName = 'With One Long Height Item';
 
-export const WithNestedScroller = () => {
-	let noScrollByWheel = boolean('noScrollByWheel', Config);
+export const WithNestedScroller = (args) => {
 	return (
 		<Scroller
 			direction="vertical"
-			focusableScrollbar={
-				prop.focusableScrollbarOption[
-					select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-				]
-			}
-			key={select('scrollMode', prop.scrollModeOption, Config) + '1'}
+			focusableScrollbar={args['focusableScrollbar']}
+			key="nested-outer"
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
 			verticalScrollbar="visible"
 		>
 			<Scroller
 				direction="horizontal"
-				focusableScrollbar={
-					prop.focusableScrollbarOption[
-						select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-					]
-				}
+				focusableScrollbar={args['focusableScrollbar']}
 				horizontalScrollbar="visible"
-				key={select('scrollMode', prop.scrollModeOption, Config) + '2'}
-				noScrollByWheel={noScrollByWheel}
+				key="nested-inner-1"
+				noScrollByWheel={args['noScrollByWheel']}
 				onKeyDown={action('onKeyDown (Nested 1st Scroller)')}
 				onScrollStart={action('onScrollStart (Nested 1st Scroller)')}
 				onScrollStop={action('onScrollStop (Nested 1st Scroller)')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
 				style={{
 					height: 'auto',
 					width: '90%'
@@ -530,23 +556,18 @@ export const WithNestedScroller = () => {
 					</Item>
 					<br />
 					<br />
-					<Item>To set or unset noScrollByWheel prop, click KNOBS below.</Item>
+					<Item>To set or unset noScrollByWheel prop, click CONTROLS below.</Item>
 				</div>
 			</Scroller>
 			<Scroller
 				direction="horizontal"
-				focusableScrollbar={
-					prop.focusableScrollbarOption[
-						select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)
-					]
-				}
+				focusableScrollbar={args['focusableScrollbar']}
 				horizontalScrollbar="visible"
-				key={select('scrollMode', prop.scrollModeOption, Config) + '3'}
-				noScrollByWheel={noScrollByWheel}
+				key="nested-inner-2"
+				noScrollByWheel={args['noScrollByWheel']}
 				onKeyDown={action('onKeyDown (Nested 2nd Scroller)')}
 				onScrollStart={action('onScrollStart (Nested 2nd Scroller)')}
 				onScrollStop={action('onScrollStop (Nested 2nd Scroller)')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
 				style={{
 					height: 'auto',
 					width: '90%'
@@ -573,25 +594,28 @@ export const WithNestedScroller = () => {
 					</Item>
 					<br />
 					<br />
-					<Item>To set or unset noScrollByWheel prop, click KNOBS below.</Item>
+					<Item>To set or unset noScrollByWheel prop, click CONTROLS below.</Item>
 				</div>
 			</Scroller>
 		</Scroller>
 	);
 };
 
+select('focusableScrollbar', WithNestedScroller, prop.focusableScrollbarOption, Config);
+boolean('noScrollByWheel', WithNestedScroller, Config);
+
 WithNestedScroller.storyName = 'With Nested Scroller';
 
-export const WithCustomizedStyle = () => (
+export const WithCustomizedStyle = (args) => (
 	<div>
 		<Scroller
-			focusableScrollbar={prop.focusableScrollbarOption[select('focusableScrollbar', ['false', 'true', '"byEnter"'], Config)]}
-			key={select('scrollMode', prop.scrollModeOption, Config)}
+			focusableScrollbar={args['focusableScrollbar']}
+			key={args['scrollMode']}
 			onKeyDown={action('onKeyDown')}
 			onScrollStart={action('onScrollStart')}
 			onScrollStop={action('onScrollStop')}
 			scrollbarTrackCss={css}
-			scrollMode={select('scrollMode', prop.scrollModeOption, Config)}
+			scrollMode={args['scrollMode']}
 			style={{height: ri.scaleToRem(804)}}
 		>
 			<div style={{height: ri.scaleToRem(1200)}}>
@@ -602,4 +626,41 @@ export const WithCustomizedStyle = () => (
 	</div>
 );
 
+select('focusableScrollbar', WithCustomizedStyle, prop.focusableScrollbarOption, Config);
+select('scrollMode', WithCustomizedStyle, prop.scrollModeOption, Config);
+
 WithCustomizedStyle.storyName = 'With Customized Style';
+
+export const WithLongContents = (args) => {
+	const content = `
+	The goal of Enact is to provide the building blocks for creating robust and maintainable applications. To that end, we’ve pulled together the best solutions for internationalization (i18n), accessibility (a11y), focus management, linting, testing and building. Then, we created a set of reusable components and behaviors on top of that. We combined these pieces and ensured that they work together seamlessly, allowing developers to focus on implementation.
+	Easy to Use
+	Enact builds atop the excellent React library, and provides a full framework to the developer. The recent boom of web technologies and related tools has led to a plethora of options available. In fact, getting started might be the most difficult part of building a modern web application.
+	Performant
+	Beyond initial setup, Enact continues to provide benefits. It was built with performance in mind, and conscious decisions were made to ensure that applications remain performant as they grow in size and complexity. This ranges from the way components are rendered to how data flows through application.
+	Customizable
+	Enact has a full set of customizable widgets that can be tuned and tweaked to the particular style of each project. Using our experience in building full UI libraries for a broad swath of devices ranging from TVs to watches, we have created a widget library whose components can easily be composed to create complex views and applications.
+	Adaptable
+	Enact was designed to produce native quality applications for a wide variety embedded web platforms. Read about Enact’s use cases and how it helps solve problems for Automotive, Robotics, TV and more.
+	`;
+	const longContents = content.repeat(50);
+
+	return (
+		<Scroller
+			focusableScrollbar={args['focusableScrollbar']}
+			key={args['scrollMode']}
+			onKeyDown={action('onKeyDown')}
+			onScrollStart={action('onScrollStart')}
+			onScrollStop={action('onScrollStop')}
+		>
+			<BodyText style={{whiteSpace: 'pre-line'}}>
+				{longContents}
+			</BodyText>
+		</Scroller>
+	);
+};
+
+select('focusableScrollbar', WithLongContents, prop.focusableScrollbarOption, Config, 'byEnter');
+select('scrollMode', WithLongContents, prop.scrollModeOption, Config);
+
+WithLongContents.storyName = 'With Long Contents';