+++ samples/sampler/stories/qa/Spotlight.js
@@ -1,5 +1,5 @@
 import {action} from '@enact/storybook-utils/addons/actions';
-import {boolean, select} from '@enact/storybook-utils/addons/knobs';
+import {boolean, select} from '@enact/storybook-utils/addons/controls';
 import Spotlight from '@enact/spotlight';
 import Pause from '@enact/spotlight/Pause';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
@@ -222,7 +222,7 @@ class PopupFocusTest extends Component {
 					return to the button used to originally open the popup. Verify this behavior for each of
 					the buttons.
 				</p>
-				<p>Use the knobs to verify 5-way behavior under different Popup configurations.</p>
+				<p>Use the controls to verify 5-way behavior under different Popup configurations.</p>
 				<Button onClick={this.handleOpenPopup}>Open Popup</Button>
 				<Button onClick={this.handleOpenPopup}>Open Popup</Button>
 				<Popup
@@ -317,6 +317,11 @@ export const CheckRerender = () => (
 );
 
 CheckRerender.storyName = 'Check Re-render';
+CheckRerender.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
 export const MultipleButtons = () => (
 	<Row align="center space-evenly">
@@ -331,6 +336,12 @@ export const MultipleButtons = () => (
 		</Cell>
 	</Row>
 );
+
+MultipleButtons.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
 export const MultipleContainers = () => (
 	<Scroller>
@@ -362,6 +373,12 @@ export const MultipleContainers = () => (
 	</Scroller>
 );
 
+MultipleContainers.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
+
 export const NestedContainers = () => (
 	<div>
 		<p>
@@ -381,6 +398,12 @@ export const NestedContainers = () => (
 	</div>
 );
 
+NestedContainers.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
+
 export const DirectionalEvents = () => (
 	<div>
 		<p>
@@ -399,28 +422,61 @@ export const DirectionalEvents = () => (
 	</div>
 );
 
+DirectionalEvents.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
+
+
 export const DisappearingSpottable = () => <DisappearTest />;
+
+DisappearingSpottable.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
 export const DisabledOnClick = () => <DisableOnClick />;
 
 DisabledOnClick.storyName = 'Disabled on Click';
+DisabledOnClick.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
 export const DisabledWithPause = () => <DisableTest />;
 
 DisabledWithPause.storyName = 'Disabled with Pause';
+DisabledWithPause.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
-export const PopupNavigation = () => (
+export const PopupNavigation = (args) => (
 	<PopupFocusTest
-		noAnimation={boolean('noAnimation', Popup, false)}
-		noAutoDismiss={boolean('noAutoDismiss', Popup, false)}
-		scrimType={select('scrimType', ['none', 'transparent', 'translucent'], Popup, 'translucent')}
-		spotlightRestrict={select('spotlightRestrict', ['self-first', 'self-only'], Popup, 'self-only')}
+		noAnimation={args['noAnimation']}
+		noAutoDismiss={args['noAutoDismiss']}
+		scrimType={args['scrimType']}
+		spotlightRestrict={args['spotlightRestrict']}
 	/>
 );
+
+boolean('noAnimation', PopupNavigation, Popup, false);
+boolean('noAutoDismiss', PopupNavigation, Popup, false);
+select('scrimType', PopupNavigation, ['none', 'transparent', 'translucent'], Popup, 'translucent');
+select('spotlightRestrict', PopupNavigation, ['self-first', 'self-only'], Popup, 'self-only');
 
 export const _FocusedAndDisabled = () => <FocusedAndDisabled />;
 
 _FocusedAndDisabled.storyName = 'Focused and Disabled';
+_FocusedAndDisabled.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
 export const NavigatingIntoOverflowContainers = () => (
 	<div>
@@ -439,15 +495,20 @@ export const NavigatingIntoOverflowContainers = () => (
 );
 
 NavigatingIntoOverflowContainers.storyName = 'Navigating into overflow containers';
+NavigatingIntoOverflowContainers.parameters = {
+	controls: {
+		hideNoControlsWarning: true
+	}
+};
 
-export const KitchenSink = () => (
+export const KitchenSink = (args) => (
 	<Column>
 		<Cell component="p" shrink>
-			Use the knobs to test the available behaviors for the spottable components below.
+			Use the controls to test the available behaviors for the spottable components below.
 		</Cell>
 		<Cell
 			component={Container}
-			spotlightDisabled={boolean('Container spotlightDisabled', Container, false)}
+			spotlightDisabled={args['Container spotlightDisabled']}
 		>
 			<Row style={{height: '100%'}}>
 				<Cell>
@@ -462,7 +523,7 @@ export const KitchenSink = () => (
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								>
 									Button
 								</Button>
@@ -474,7 +535,7 @@ export const KitchenSink = () => (
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								>
 									Transparent
 								</Button>
@@ -486,14 +547,14 @@ export const KitchenSink = () => (
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								/>
 								<Input
 									onSpotlightDown={action('onSpotlightDown')}
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								/>
 							</div>
 							<div>
@@ -502,7 +563,7 @@ export const KitchenSink = () => (
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								>
 									{Items}
 								</Picker>
@@ -512,7 +573,7 @@ export const KitchenSink = () => (
 									onSpotlightLeft={action('onSpotlightLeft')}
 									onSpotlightRight={action('onSpotlightRight')}
 									onSpotlightUp={action('onSpotlightUp')}
-									spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+									spotlightDisabled={args['Spottable spotlightDisabled']}
 								>
 									{Items}
 								</Picker>
@@ -522,14 +583,14 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							/>
 							<Item
 								onSpotlightDown={action('onSpotlightDown')}
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								Item
 							</Item>
@@ -539,7 +600,7 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								Item with label
 							</Item>
@@ -548,7 +609,7 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								CheckboxItem
 							</CheckboxItem>
@@ -557,7 +618,7 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								FormCheckboxItem
 							</FormCheckboxItem>
@@ -566,7 +627,7 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								RadioItem
 							</RadioItem>
@@ -575,7 +636,7 @@ export const KitchenSink = () => (
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
 								onSpotlightUp={action('onSpotlightUp')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 							>
 								SwitchItem
 							</SwitchItem>
@@ -591,13 +652,13 @@ export const KitchenSink = () => (
 							<DatePicker
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 								title="DatePicker"
 							/>
 							<TimePicker
 								onSpotlightLeft={action('onSpotlightLeft')}
 								onSpotlightRight={action('onSpotlightRight')}
-								spotlightDisabled={boolean('Spottable spotlightDisabled', Container, false)}
+								spotlightDisabled={args['Spottable spotlightDisabled']}
 								title="TimePicker"
 							/>
 						</Cell>
@@ -607,3 +668,6 @@ export const KitchenSink = () => (
 		</Cell>
 	</Column>
 );
+
+boolean('Container spotlightDisabled', KitchenSink, Container, false);
+boolean('Spottable spotlightDisabled', KitchenSink, Container, false);